tests/submission/Ms2ger/the-elements-of-html/scripting/the-script-element/script-languages-02.html

<!DOCTYPE html>
<title>Script @type: JavaScript types</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="http://www.whatwg.org/html5/#scriptingLanguages">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
function testAttribute(attr, val) {
  test(function() {
    assert_false(ran, "ran variable not reset");
    var script = document.createElement("script");
    script.setAttribute(attr, val);
    script.textContent = "ran = true;"
    document.body.appendChild(script);
    assert_true(ran);
    ran = false
  }, "Script should run with " + attr + "=" + val);
}
function testType(type) {
  testAttribute("type", type);
}
function testLanguage(lang) {
  testAttribute("language", lang);
}
var application = [
  "ecmascript",
  "javascript",
  "x-ecmascript",
  "x-javascript"
];
var text = [
  "ecmascript",
  "javascript",
  "javascript1.0",
  "javascript1.1",
  "javascript1.2",
  "javascript1.3",
  "javascript1.4",
  "javascript1.5",
  "jscript",
  "livescript",
  "x-ecmascript",
  "x-javascript"
];
var ran = false;
application.map(function(t) { return "application/" + t; }).forEach(testType);
application.map(function(t) { return ("application/" + t).toUpperCase(); }).forEach(testType);
text.map(function(t) { return "text/" + t; }).forEach(testType);
text.map(function(t) { return ("text/" + t).toUpperCase(); }).forEach(testType);

text.forEach(testLanguage);
text.map(function(t) { return t.toUpperCase(); }).forEach(testLanguage);
</script>
